Fix behavior of indicator levels
An indicator can be associated to plans through the IndicatorLevel
model. When listing indicators, the user can choose to see either (a)
all indicators for the organization behind the user's active admin plan,
or (b) only those that are also in the user's active admin plan.

When editing an indicator associated to the active admin plan, the user
can delete the association with this plan by selecting "[not in active
plan]" as the indicator's level.

When creating an indicator or editing an indicator not associated to the
active admin plan, the user can create such an association by selecting
a proper indicator level.