Add decorator to override methods in types generated by Grapple
When you have a ListBlock in a StructBlock in a StreamField (as we have
in `qa_section` of the `body` of `StaticPage`) and the list is empty,
Grapple by default resolves the field of the list list to `None` instead
of `[]`. This might be a bug in Grapple. As a workaround, the proposed
decorator can be used to selectively override resolvers where this
behavior is undesired.