Fix docker-compose.yml
If APP_IMAGE or DB_IMAGGE are not set, docker-compose will complain
about the blank value for the image option.

Publishing ports with the "ports" option is quite deployment-specific
and should better be done in an override file. Some deployments may
choose not to publish ports, for example because they use a reverse
proxy in a Docker container in the same network.