Work around the fact that modeltrans assumes a single main language
In the main tab (i.e., not a tab holding only translations of i18n
fields) of a tabbed modeladmin interface for a plan-related model, we
want to display i18n fields in the language of the primary language of
the plan. (For models not related to a plan, we arguably want to display
these fields in the currently activated language.) This is tricky since
modeltrans always assumes a translatable field contains a value in
Django's default language and all translations of the field are in the
i18n field. For plan-related models where, for example, the plan does
not even support Django's default language, such fields must remain
empty. Unfortunately we must therefore remove database constraints that
such fields have a value.