Replace modeltrans with our own variant in new app i18n
This allows us to choose the default language per instance, so the
original translatable field can be in a language other than Django's
default.

Add primary_language field to models that have no primary language
specified (or derivable by following foreign keys) yet. This is
necessary so that we know what language the original translatable fields
are in.