Get rid of TranslatedModelMixin in favor of modeltrans fallback
Turns out modeltrans can be made to work with our assumption that
the original translatable fields are in varying non-default languages in
general. For this, we just need to set the fallback languages to the
empty list. This will give us the original translatable field as a
fallback.

So I put some code from modeltrans related to fallback, that I had
originally removed, back in.