Harmonize icons of categories and common categories
We allow SVGs and raster images for both models. We rely on the
wagtailsvg package until Wagtail gets built-in SVG support.

Since the way SVGs are stored in the DB has changed, the migration will
raise an error if there are CategoryImage instances with SVG data.
Migration must be done manually at this point.