Remove recently introduced default values for environment variables
These values interfere with our overrides. For example, if we set
`POSTGRES_PASSWORD=${POSTGRES_PASSWORD:-watchpass}` for the DB service in
docker-compose.yml, then both `POSTGRES_PASSWORD` and
`POSTGRES_PASSWORD_FILE` will be set. The former because the default
`watchpass` will kick in; the latter because we set it to the Docker
config in our override file. The postgres image aborts with an error if
both are specified.

Moreover, it may potentially confusing for us if we inspect a service,
find that, e.g., for the Watch app service, `DATABASE_URL` is set to
some value, but it's actually ignored because we happen to also set
`ENV_FILE`, which sets a different value. This can easily be overlooked.