Add setting field in Plan model to set action staleness limit
It is currently used for sending notifications when
actions have not been updated in N days.

Will be used in UIs elsewhere too.

The field is allowed to be null in the database for a plan, in which
case the default value (defined in code) is used for the plan.