Remove interpolation parameter from help_text
Somewhere in the depths of wagtail, the help_text gets eagerly
evaluated. This is apparently done in a context where DEFAULT_LANGUAGE
is used as the language. Maybe it gets translated again or something
but the string won't match with the interpolation already done.
I'm unwilling to debug further at this point.