Fix reordering of FixedSlugPage instances
This didn't work because Wagtail checks whether the page is allowed as a
child of the new parent even if this didn't change. We used to specify
that there are no allowed parent page types for a FixedSlugPage in order
to prevent users from creating such pages manually. This fixes the
parent page type definition and prevents manual creation in a different
way.