Allow deactivating users
When deleting a person from wagtail admin, the corresponding user is
deactivated with a record of who did this an when. Currently only
superusers can do this.

The user is kept in the database to persist the foreign keys to
models, revisisions, etc.

A separate maintenance task will permanently delete the users when
enough time has passed.