Use bulk_update in REST API to save order for non-Treebeard models
Otherwise saving, e.g.,  a bunch of actions with a single PUT request
will cause a quadratic number of calls to `Action.save()`: For each
updated action, the order of all other actions is updated.

We now call `bulk_update()` once for each action in the PUT request. We
are thus still left with a linear number of calls to `save()`, which
probably should be improved to a single bulk update at some point.