Normalize indicator goals as well
We might need to change the calculation logic at some point. Right now,
when normalizing by population, for example, we use the goal of the
population indicator for a certain year, whereas perhaps we should use a
value (as opposed to a goal) even if it's in the future. Maybe add some
way of distinguishing actual values from forecast values.