Switch to using ManifestStaticFileStorage for cache busting
The system adds checksums to filenames when collecting static files.

Only works on production with DEBUG=False.

Don't use the dev environment manage.py runserver with DEBUG=False because static files
do not work at all with that setup.

The reason for taking this into use was that we were earlier relying
on Wagtail's similar mechanism, but it turned out that it only took into
account new Wagtail releases, meaning the cache busting checksum
remained the same as long as the Wagtail version was the same.