Always use cached instances to fix deferred operations with dependencies
If we have several deferred operations, the later ones may depend on the
state established by the earlier ones. We thus need to make sure that we
keep that state up to date in the cached instances, and to always use
the cached instances as opposed to stuff that's coming from our DB.

I put the initialization of the cache into
`ActionSerializer.__init__()`, which I hope solves these issues.