Support multilingual read_only fields
The read_only AttributeFieldPanel needs to be aware of the language
and use it when rendering the field contents to text.

Also render OptionalChoiceWithText read_only attributes better,
with multilingual support. Instead of two read_only panels, add just one.
(Otherwise the the attribute contents via __str__ is rendered twice)