Change reference to plan before saving AttributeType
It's not enough to do it afterwards with `UpdateReferencesVisitor`. This
is because the `AutoSlugField` for `identifier` in `AttributeType` has
`always_update=True` and is unique per plan. So when saving the copy, it
references the same plan as the original, thus getting a suffix appended
to the slug to make it unique.