Reimplement dashboard excel export ActionResponsibleParty logic
Previously, the ActionResponsibleParty only showed one organization
in reports, the first one with the "primary" role. Now, we display
all of the organizations in 3 different columns according to role.

Special care is taken when calculating the parent, that still
uses the "first" org based on some criteria. The parent is needed
for certain report summaries where figures are grouped by organization.