Update cluster related objects in plan copy
This should fix an error where copying a plan with drafts and then
publishing a draft in the copy would steal related objects (e.g.,
contact persons) from the original action. That would happen because the
serialized data in the revisions still contains references to instances
belonging to the original plan.