Fix references in drafts after deleting action-related instances
When deleting an instance related to a model, such as a contact person,
drafts of that action may reference the deleted instance, which causes
an error when trying to publish the action. Here we remove the reference
from the revision content so that, when the draft is published, the
related instance is created anew instead of trying (and failing) to
change the one that doesn't exist anymore.