Add flag exclude_from_search to Plan model
This can be used to handle cases where a plan is public but should still
not show up in search results, for example when checking "search only
across other cities' plans". This could make sense for plans that are
for demo purposes, or for plans that are, albeit published and available
to anyone who knows the URL, still used somewhat internally by a customer.

[Asana task](https://app.asana.com/0/1206017643443542/1209246928297785)