Clean up test mode mutations and avoid duplicating types
We only return PKs of created objects for now. Once we migrated the
GraphQL types for our models from Graphene to Strawberry, we could use
the Strawberry types for returning the actual instances.

I removed the test mode mutation for creating organizations because
there is already a mutation in the `orgs` app. The test mode mutation is
therefore no longer necessary.