Refactor common category admin edit handler
Refactor common category admin edit handler to not use the request or
instance, as those are not available when migrating out of modeladmin.

The issue mentioned in the comment about invalidated icons seems to not
exist anymore, so remove the comment and move to CondensedInlinePanel.

Instead of translation tabs, handle translations using
TranslatedFieldPanel.

As editing common categories is nowadays only allowed for superusers
anyway, do not worry about the individual panel permissions.