Use color picker panel where appropriate
Caveat: The widget we use only supports hex colors. For some models, our
database currently contains non-hex CSS colors, such as "red" or
"rgb(255, 0, 0)". I did not bother writing a data migration for
converting all such values to hex. When users try to save an existing
model instance with a non-hex color, they will get an error. But I
suppose we can live with that since these cases are few, so perhaps we
can expect users to convert to hex themselves once that happens, or send
a support request.

Note that `ActionStatus` and `ActionImplementationPhase` use key codes
such as "red030" that seem to refer to the UI theme. In these cases, a
color picker is not appropriate, so these panels continue to be plain
text.

I took this opportunity to disallow NULL for the color values, which may
also be empty strings, because allowing both NULL and empty values
[is discouraged](https://docs.djangoproject.com/en/5.2/ref/models/fields/#django.db.models.Field.null)
unless it cannot be avoided (for example in the presence of uniqueness
constraints).