Save related actions when saving action (published or draft)
This used to be broken. When creating an action, related actions could
be set, but after creation, any attempt to edit the related actions
would be ignored.

This is because the relationship between an action and the related
action is not "parental" (we use `ManyToManyField`) and there is no code
that puts this relation into the revision.

While investigating this issue, I realized that other relationships
are in principle also affected. For example, the categories of an
action. There, however, we made the questionable decision to set the
categories of the `Action` instance persisted in the database even when
only saving a draft. That is, the published action's categories are
affected when clicking "save draft".

This commit "fixes" the bug that changes to the related actions are
ignored by handling it in the same way as categories. Note that now for
related actions we have the same arguably buggy behavior as we have for
action categories. However, I believe it is better to have one bug
instead of two. We can, and should, fix the other bug as soon as
possible.