POC: Preserve published action when changing related actions in draft
This makes `RelatedActionsThrough` an inline model of `Action` and
replaces the autocomplete field for related actions by an inline panel
-- a `MultipleChooserPanel` to be precise.

This fixes the buggy behavior that changing related actions and clicking
"save draft" updates the related actions of the published action, not
just the draft. In fact, action drafts previously did not contain any
information about related actions but took that information straight
from the published version.

While this commit may fix the logic, it arguably makes the UX worse --
that is, if we disregard the fact that changing a published action
without actually wanting to do so is pretty bad UX to begin with. Also,
the current excess of headings and labels of the inline panel should be
improved.

As another task to do in the future, the new `MultipleChooserPanel` is
not customizable like many other fields are by means of the
`CustomizableBuiltInFieldPanel`. The autocomplete version used
`CustomizableBuiltInFieldPanel`, but I have not verified if this used to
work.

So far, this is a proof-of-concept. The problem persists for other
many-to-many relationships as well; for example, action categories. A
similar fix still would need to be applied to those cases.