Copy indicator category assignment when copying plans
We designed the plan copying feature so that indicators are not copied.
They are supposed to be shared between the original plan and the copy.
While categories themselves were copied, the association between
indicators and the copies of their categories was not made, which caused
some issues.

This fixes that. We may want to optionally copy indicators too, even
though we'll have to be careful due to them being shared between plans
in general.