Fix issues with attribute visibility in our APIs
The GraphQL field `actionAttributeTypes` in `PlanNode` returned all
attribute types, even those whose instances are not visible for the
user. In order to avoid showing a column for these attribute types in,
e.g., the action grid editor, This commit removes them from
`actionAttributeTypes`, which I think is better than filtering in the
UI.

In the REST API, our attribute serializers included invisible
attributes, which leaks information. This commit fixes that.